Move PID logic to a dedicated class
Helps integrating Auto tune code and removes the burden and knowledge
from other components.
This is a first step as every PID part from the Roaster class have been
removed and replaced with call to PID functions.
Some of these calls are redundant, and we might just expose the PID to
other classes that need to do operation on it, removing intermediary
calls.